+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 - Update code to use Java 7
 - Mac OSX app updated to support Java 7 and no longer use Java 6. Users must redownload the app
 - Added in Download Clearer Tool to remove all your downloads from downloading packs to clear space if needed
+- When a direct download is used for a mod, check the content length from the url to see if it should be re downloaded
 - Fixed issue that gave a bad '32 bit Java on 64 bit Windows' warning when 64 bit Java was installed in the x86 directory